Finished Notification, map permissions working and reset challenge when claim reward
